# Conflicts:
#	package/commerce/src/OroB2B/Bundle/ProductBundle/Resources/views/layouts/default/orob2b_product_frontend_product_index/product_index_row.html.twig
#	package/commerce/src/OroB2B/Bundle/ProductBundle/Resources/views/layouts/default/orob2b_product_frontend_product_index/product_index_row.yml
#	package/commerce/src/OroB2B/Bundle/ShoppingListBundle/Layout/DataProvider/FrontendShoppingListProductUnitsQuantityDataProvider.php
#	package/commerce/src/OroB2B/Bundle/ShoppingListBundle/Resources/views/layouts/default/imports/oro_product_list_item/shopping_list_buttons.yml
#	package/commerce/src/OroB2B/Bundle/ShoppingListBundle/Tests/Functional/Entity/Repository/LineItemRepositoryTest.php
#	package/commerce/src/OroB2B/Bundle/ShoppingListBundle/Tests/Unit/Layout/DataProvider/FrontendShoppingListProductUnitsQuantityDataProviderTest.php

+++ src/Oro/Bundle/WindowsBundle/Resources/public/js/widget/dialog-widget.js
@@ -26,7 +26,8 @@ define(function(require) {
             type: 'dialog',
             dialogOptions: null,
             stateEnabled: true,
-            incrementalPosition: true
+            incrementalPosition: true,
+            preventModelRemoval: false
         }),
 
         // Windows manager global variables
@@ -139,7 +140,7 @@ define(function(require) {
                 return;
             }
             dialogManager.remove(this);
-            if (this.model) {
+            if (this.model && !this.options.preventModelRemoval) {
                 this.model.destroy({
                     error: _.bind(function(model, xhr) {
                         // Suppress error if it's 404 response and not debug mode
@@ -273,6 +274,7 @@ define(function(require) {
 
         _clearActionsContainer: function() {
             this.widget.dialog('actionsContainer').empty();
+            this.actionsEl = null;
         },
 
         _renderActions: function() {
